update board when the board state changes
I created a function updateBoard, and subscribed it to the redux store.
I'm removing the board, creating a new one and adding the new one to the
page. This is probably the least efficient way of doing it, but we'll
see how it goes. If I can make this work maybe I'll re-make it with
React again to get the rendering efficiency.

I didn't write tests for this code, as I wasn't sure how to test the
rendered state of the page outside of React. I'll look that up for next
time.